+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @supersubset/theme
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @supersubset/schema@0.1.2
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @supersubset/schema@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes